# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: string;
  address: string;
}

export type ContactStage = typeof ContactStage[keyof typeof ContactStage];


export const ContactStage = {
  lead: 'lead',
  prospect: 'prospect',
  client: 'client',
  purchaser: 'purchaser',
} as const;

export interface Contact {
  id: number;
  name: string;
  phone: string;
  email: string;
  notes?: string;
  properties?: Property[];
  stage: ContactStage;
  createdAt: string;
  updatedAt: string;
}

export type InsertContactStage = typeof InsertContactStage[keyof typeof InsertContactStage];


export const InsertContactStage = {
  lead: 'lead',
  prospect: 'prospect',
  client: 'client',
  purchaser: 'purchaser',
} as const;

export interface InsertContact {
  name: string;
  phone: string;
  email: string;
  notes?: string;
  properties?: Property[];
  stage?: InsertContactStage;
}

export type UpdateContactStage = typeof UpdateContactStage[keyof typeof UpdateContactStage];


export const UpdateContactStage = {
  lead: 'lead',
  prospect: 'prospect',
  client: 'client',
  purchaser: 'purchaser',
} as const;

export interface UpdateContact {
  name?: string;
  phone?: string;
  email?: string;
  notes?: string;
  properties?: Property[];
  stage?: UpdateContactStage;
}

export interface PersonalContact {
  id: number;
  name: string;
  phone: string;
  email: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface InsertPersonalContact {
  name: string;
  phone: string;
  email: string;
  notes?: string;
}

export interface UpdatePersonalContact {
  name?: string;
  phone?: string;
  email?: string;
  notes?: string;
}

export interface Error {
  error: string;
}

export type DeleteContact200 = {
  message?: string;
  contact?: Contact;
};