# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactStage } from './contactStage';
import type { Property } from './property';

export interface Contact {
  id: number;
  name: string;
  phone: string;
  email: string;
  notes?: string;
  properties?: Property[];
  stage: ContactStage;
  createdAt: Date;
  updatedAt: Date;
}