# lib/api-zod/src/generated/types/insertContact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InsertContactStage } from './insertContactStage';
import type { Property } from './property';

export interface InsertContact {
  name: string;
  phone: string;
  email: string;
  notes?: string;
  properties?: Property[];
  stage?: InsertContactStage;
}