# lib/api-zod/src/generated/types/updateContact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Property } from './property';
import type { UpdateContactStage } from './updateContactStage';

export interface UpdateContact {
  name?: string;
  phone?: string;
  email?: string;
  notes?: string;
  properties?: Property[];
  stage?: UpdateContactStage;
}